Update SQL (#7)
* bug fix and added color themes

* extensions are enabled on conf.ini now

* new poll and optimized role poll

* modified help menu

* Implemented pretty_help and polls

* Develop (#1)

* modified help menu

* Implemented pretty_help and polls

* new commands and bugfix of polls. Now admin privs are required

* first draft of permissions

* basic permission system

* decorator for checking custom permission

* added custom permissions and cooldowns to polls

* part implementation of role poll modification

* Refactoring and new logging

* Merge branch 'update' of github.com:MysticBanana/GeoffreyButler into develop

# Conflicts:
#	cogs/__init__.py
#	core/bot.py
#	data/models/permission.py

* Rebuild data management. Interacting with a sqlite db now. WIP

* New logging - Creates new logging class with multiple handlers which got different log levels. Also moved the output of the discord logger to the same file

* New logging - Activity Plugin

* Activity Plugin - added status command

+++ core/helper/interactive_menu.py
@@ -81,7 +81,7 @@ def request_decorate(func):
         bot_msg = await bot.responses.send(channel=channel, make_embed=False, content=content)
 
         msg = await bot.wait_for("message", check=lambda m: (m.author == user and m.channel == channel),
-                                timeout=40.0)
+                                timeout=65.0)
 
         if msg.content == ignore_character:
             return None
@@ -121,7 +121,7 @@ async def request_emoji(bot, channel, user: discord.Member, content, counter: in
     bot_msg = await bot.responses.send(channel=channel, make_embed=False, content=content)
 
     msg: discord.Message = await bot.wait_for("message", check=lambda m: (m.author == user and m.channel == channel),
-                                  timeout=40.0)
+                                  timeout=65.0)
     content = msg.content
     await bot_msg.delete()
     await msg.delete()
@@ -142,22 +142,25 @@ def request_bool(content) -> bool:
 
 
 async def request_roles(bot, channel, user: discord.Member, content, counter: int = 0) -> List[discord.Role]:
-    bot_msg = await bot.responses.send(channel=channel, make_embed=False, content=content)
+    bot_msg = await bot.responses.send(channel=channel, make_embed=False, content=f"{content}\n*Type"
+                                                                                  f" <cancel> to cancel the request*")
 
     if counter > 3:
         await bot.responses.send(channel=channel, make_embed=False, content="Maximum tries")
         return []
 
     msg = await bot.wait_for("message", check=lambda m: (m.author == user and m.channel == channel),
-                             timeout=40.0)
-
+                             timeout=65.0)
     content = msg.content
+
+    if content == "cancel":
+        return []
+
     roles = msg.role_mentions
     await bot_msg.delete()
     await msg.delete()
 
     if len(roles) == 0:
-        # retry
         return await request_roles(bot, channel, user, content, counter + 1)
 
     return roles
@@ -171,7 +174,7 @@ async def request_channel(bot, channel, user: discord.Member, content, counter: 
     bot_msg = await bot.responses.send(channel=channel, make_embed=False, content=content)
 
     msg: discord.Message = await bot.wait_for("message", check=lambda m: (m.author == user and m.channel == channel),
-                             timeout=40.0)
+                             timeout=65.0)
 
     content = msg.content
     channels = msg.channel_mentions
